requirements.txt
fastapi
uvicorn
torch
openai-whisper
pdf2image
pytesseract
rapidfuzz
numpy
pillow
